Order status

+++ src/services/orderService.js
@@ -11,10 +11,15 @@ import { sellerModel } from '~/models/sellerModel'
 // import { sellerModel } from '~/models/sellerModel'
 import { shopModel } from '~/models/shopModel'
 import { GHNProvider } from '~/providers/GHNProvider'
+import { getNameDetailForGhn } from '~/utils/algorithms'
 // import { GHNProvider } from '~/providers/GHNProvider'
 import ApiError from '~/utils/ApiError'
 import { ORDER_STATUS } from '~/utils/constants'
 
+/**
+ * Buyer APIs
+ * @author taiki and levi
+ */
 const clusterOrder = async (buyerId, reqBody) => {
   try {
     const itemList = reqBody.itemList
@@ -26,26 +31,38 @@ const clusterOrder = async (buyerId, reqBody) => {
       const { productId, typeId, quantity } = item
       const product = await productModel.findOneById(productId)
 
-      if (!product) throw new ApiError(StatusCodes.NOT_ACCEPTABLE, 'Sản phẩm không hợp lệ')
+      if (!product)
+        throw new ApiError(StatusCodes.NOT_ACCEPTABLE, 'Sản phẩm không hợp lệ')
 
       // Thông số của loại sản phẩm
-      if (!product.typeFeature) throw new ApiError(StatusCodes.CONFLICT, 'Sản phẩm không có type')
-      const { typeName, discount, price } = (product.typeFeature).find((type) => type.typeId.toString() === typeId)
+      if (!product.typeFeature)
+        throw new ApiError(StatusCodes.CONFLICT, 'Sản phẩm không có type')
+      const { typeName, discount, price } = product.typeFeature.find(
+        (type) => type.typeId.toString() === typeId
+      )
       const finalPrice = price * (1 - discount / 100)
 
-
       // Tìm shop đầu tiên có loại sản phẩm này
-      if (!product.shopTypes) throw new ApiError(StatusCodes.CONFLICT, 'Sản phẩm không có mẫu này')
+      if (!product.shopTypes)
+        throw new ApiError(StatusCodes.CONFLICT, 'Sản phẩm không có mẫu này')
       const shopList = product.shopTypes
-      const foundShop = shopList.find((shop) => shop.types.some((type) => type.typeId.toString() === typeId && type.stock >= quantity))
+      const foundShop = shopList.find((shop) =>
+        shop.types.some(
+          (type) => type.typeId.toString() === typeId && type.stock >= quantity
+        )
+      )
 
-      if (!foundShop) throw new ApiError(StatusCodes.NOT_FOUND, 'Không tìm thấy shop có mẫu sản phẩm này')
+      if (!foundShop)
+        throw new ApiError(
+          StatusCodes.NOT_FOUND,
+          'Không tìm thấy shop có mẫu sản phẩm này'
+        )
 
       const foundShopId = foundShop.shopId
 
-
       const shop = await shopModel.findOneById(foundShopId)
-      if (!shop) throw new ApiError(StatusCodes.NOT_ACCEPTABLE, 'Shop không hợp lệ')
+      if (!shop)
+        throw new ApiError(StatusCodes.NOT_ACCEPTABLE, 'Shop không hợp lệ')
 
       const shopAddress = shop.shopAddress
 
@@ -56,9 +73,12 @@ const clusterOrder = async (buyerId, reqBody) => {
       const lastOrderQuantity = quantity % maxQuantityPerOrder
       if (lastOrderQuantity > 0) itemQuantityArr.push(lastOrderQuantity)
 
-
       itemQuantityArr.forEach((quantity) => {
-        let foundClusteredOrder = clusteredOrderList.find((order) => order.shopId.toString() === foundShopId.toString() && order.orgPrice + finalPrice * quantity <= 50000000)
+        let foundClusteredOrder = clusteredOrderList.find(
+          (order) =>
+            order.shopId.toString() === foundShopId.toString() &&
+            order.orgPrice + finalPrice * quantity <= 50000000
+        )
         // Validate lại item trong itemList
         const validItem = {
           ...item,
@@ -83,7 +103,6 @@ const clusterOrder = async (buyerId, reqBody) => {
           clusteredOrderList.push(newClusteredOrder)
         }
       })
-
     }
     return clusteredOrderList
   } catch (error) {
@@ -91,10 +110,8 @@ const clusterOrder = async (buyerId, reqBody) => {
   }
 }
 
-
-const addOrder = async (buyerId, reqBody) => {
+const addOrder = async (buyerId, reqBody, buyNow = false) => {
   try {
-
     const orderData = {
       ...reqBody,
       buyerId: buyerId
@@ -102,14 +119,25 @@ const addOrder = async (buyerId, reqBody) => {
 
     const insertedOrder = await orderModel.addOrder(orderData)
 
-
     let updatedItemList = []
     for (let item of reqBody.itemList) {
-      const updatedItem = await productModel.increaseStock(item.productId, item.typeId, reqBody.shopId, -item.quantity)
-      await cartModel.deleteItem(buyerId, {
-        productId: item.productId,
-        typeId: item.typeId
-      })
+      // Decrease stock
+      const updatedItem = await productModel.increaseStock(
+        item.productId,
+        item.typeId,
+        reqBody.shopId,
+        -item.quantity
+      )
+
+      // Increase sold
+      await productModel.update(item.productId, { sold: updatedItem.sold + 1 })
+
+      // Với trường hợp Mua ngay, không xóa trong cart
+      if (!buyNow)
+        await cartModel.deleteItem(buyerId, {
+          productId: item.productId,
+          typeId: item.typeId
+        })
       updatedItemList.push(updatedItem)
     }
 
@@ -119,38 +147,62 @@ const addOrder = async (buyerId, reqBody) => {
   }
 }
 
-const getAllOrdersForSeller = async (sellerId) => {
+const getAllOrders = async (buyerId) => {
   try {
-    const result = await orderModel.getAllOrdersForSeller(sellerId)
+    const result = await orderModel.getAllOrders(buyerId)
     return result
-  } catch (error) { throw error }
+  } catch (error) {
+    throw error
+  }
 }
 
-const getAllOrdersForBuyer = async (buyerId) => {
+// ----------------------------------------------------------------------------
+// ----------------------------------------------------------------------------
+// ----------------------------------------------------------------------------
+
+/**
+ * Seller APIs
+ * @author taiki and levi
+ */
+const seller_getAllOrders = async (sellerId, isLatest = false) => {
   try {
-    const result = await orderModel.getAllOrdersForBuyer(buyerId)
+    const result = await orderModel.seller_getAllOrders(sellerId, isLatest)
     return result
-  } catch (error) { throw error }
+  } catch (error) {
+    throw error
+  }
 }
 
-const updateOrderStatus = async (reqBody) => {
+const seller_updateOrderStatus = async (reqBody) => {
   try {
+    let createdGHNOrder = null
     const { orderId, status } = reqBody
-    const updatedOrder = await orderModel.updateOrderStatus(orderId, status)
 
     if (status === ORDER_STATUS.SHIPPING) {
       // --------------------Xử lí API Create Order GHN---------------------------
       const order = await orderModel.findOneById(orderId)
-      if (!order) throw new ApiError(StatusCodes.NOT_ACCEPTABLE, 'Đơn hàng không hợp lệ')
-      const { sellerId, shopId, finalPrice, buyerPhone, buyerName, note, buyerAddress, shippingFee, itemList } = order
+      if (!order)
+        throw new ApiError(StatusCodes.NOT_ACCEPTABLE, 'Đơn hàng không hợp lệ')
+      const {
+        sellerId,
+        shopId,
+        finalPrice,
+        buyerPhone,
+        buyerName,
+        note,
+        shippingFee,
+        itemList
+      } = order
+
+      const buyerAddress = await getNameDetailForGhn(order.buyerAddress)
+
       const seller = await sellerModel.findOneById(sellerId)
       const sellerName = seller.name
 
       // Xử lí thông tin shop
       const shop = await shopModel.findOneById(shopId)
-      const shopAddress = shop.shopAddress
+      const shopAddress = await getNameDetailForGhn(shop.shopAddress)
       const shopPhone = shop.phone
-
 
       // Thông tin chưa được xử lí
       let insuranceValue = 0
@@ -161,12 +213,17 @@ const updateOrderStatus = async (reqBody) => {
 
       // Xử lí thông số đơn hàng
       const totalWeight = itemList.reduce((sum, item) => sum + item._weight, 0)
-      const orderLength = Math.max(itemList.map(item => item._length))
-      const orderWidth = Math.max(itemList.map(item => item._width))
+      const orderLength = Math.max(itemList.map((item) => item._length))
+      const orderWidth = Math.max(itemList.map((item) => item._width))
       const orderHeight = itemList.reduce((sum, item) => sum + item._height, 0)
 
       let GHNdata
-      if (totalWeight > 50000 || orderLength > 200 || orderWidth > 200 || orderHeight > 200) {
+      if (
+        totalWeight > 50000 ||
+        orderLength > 200 ||
+        orderWidth > 200 ||
+        orderHeight > 200
+      ) {
         // Hàng nặng
         const serviceTypeId = 5
         // Xử lí itemList
@@ -186,20 +243,20 @@ const updateOrderStatus = async (reqBody) => {
           from_name: sellerName,
           from_phone: shopPhone,
           from_address: shopAddress.address,
-          from_ward_name: shopAddress.ward,
-          from_district_name: shopAddress.district,
-          from_province_name: shopAddress.province,
+          from_ward_name: shopAddress.wardName,
+          from_district_name: shopAddress.districtName,
+          from_province_name: shopAddress.provinceName,
           to_name: buyerName,
           to_phone: buyerPhone,
           to_address: buyerAddress.address,
-          to_ward_name: buyerAddress.ward,
-          to_district_name: buyerAddress.district,
-          to_province_name: buyerAddress.province,
+          to_ward_name: buyerAddress.wardName,
+          to_district_name: buyerAddress.districtName,
+          to_province_name: buyerAddress.provinceName,
           return_phone: shopPhone,
           return_address: shopAddress.address,
-          return_district_name: shopAddress.district,
-          return_ward_name: shopAddress.ward,
-          return_province_name: shopAddress.province,
+          return_district_name: shopAddress.districtName,
+          return_ward_name: shopAddress.wardName,
+          return_province_name: shopAddress.provinceName,
           cod_amount: finalPrice - shippingFee,
           content: content,
           insurance_value: insuranceValue,
@@ -217,20 +274,20 @@ const updateOrderStatus = async (reqBody) => {
           from_name: sellerName,
           from_phone: shopPhone,
           from_address: shopAddress.address,
-          from_ward_name: shopAddress.ward,
-          from_district_name: shopAddress.district,
-          from_province_name: shopAddress.province,
+          from_ward_name: shopAddress.wardName,
+          from_district_name: shopAddress.districtName,
+          from_province_name: shopAddress.provinceName,
           to_name: buyerName,
           to_phone: buyerPhone,
           to_address: buyerAddress.address,
-          to_ward_name: buyerAddress.ward,
-          to_district_name: buyerAddress.district,
-          to_province_name: buyerAddress.province,
+          to_ward_name: buyerAddress.wardName,
+          to_district_name: buyerAddress.districtName,
+          to_province_name: buyerAddress.provinceName,
           return_phone: shopPhone,
           return_address: shopAddress.address,
-          return_district_name: shopAddress.district,
-          return_ward_name: shopAddress.ward,
-          return_province_name: shopAddress.province,
+          return_district_name: shopAddress.districtName,
+          return_ward_name: shopAddress.wardName,
+          return_province_name: shopAddress.provinceName,
           cod_amount: finalPrice - shippingFee,
           content: content,
           weight: totalWeight,
@@ -245,20 +302,31 @@ const updateOrderStatus = async (reqBody) => {
           required_note: requiredNote
         }
       }
-      const validGHNdata = await GHNProvider.GHN_ORDER_SCHEMA.validateAsync(GHNdata, { abortEarly: false })
+      const validGHNdata = await GHNProvider.GHN_ORDER_SCHEMA.validateAsync(
+        GHNdata,
+        { abortEarly: false }
+      )
 
-      const createdGHNOrder = await createGHNOrder(validGHNdata)
-      return { updatedOrder, createdGHNOrder }
+      createdGHNOrder = await createGHNOrder(validGHNdata)
     }
 
-    return updatedOrder
-  } catch (error) { throw error }
+    const updatedOrder = await orderModel.seller_updateOrderStatus(
+      orderId,
+      status
+    )
+    return { updatedOrder, createdGHNOrder }
+  } catch (error) {
+    throw error
+  }
 }
 
 export const orderService = {
+  // Buyer
   addOrder,
   clusterOrder,
-  getAllOrdersForSeller,
-  updateOrderStatus,
-  getAllOrdersForBuyer
+  getAllOrders,
+
+  // Seller
+  seller_getAllOrders,
+  seller_updateOrderStatus
 }
